+++ app/CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to the penguin **CLI app** will be documented here.
 ## [Unreleased]
 ### Added
 - Add `--open` flag to open the browser automatically
+- Mounted file system paths are now automatically watched for file changes. The
+  browser sessions will reload automatically if anything changes.
+- Add `--no-auto-watch` flag to disable auto watch behavior.
 
 
 ## 0.1.0 - 2021-03-03